+++ include/conslr/theme.hpp
@@ -14,16 +14,16 @@ namespace conslr
 {
     struct Theme
     {
-        int32_t borderHorizontal = '-'; //!<Horizontal border character
-        int32_t borderVertical = '|'; //!<Vertical border character
-        int32_t borderCornerTl = '+'; //!<Top left corner border character
-        int32_t borderCornerTr = '+'; //!<Top right corner border character
-        int32_t borderCornerBl = '+'; //!<Bottom left corner border character
-        int32_t borderCornerBr = '+'; //!<Bottom right corner border character
-        int32_t scrollbarCharacter = (11 * 16) + 10; //!<Scrollbar character
-        int32_t scrollbarCharacterHorizontal = (12 * 16) + 13; //!<Horizontal scrollbar character
-        int32_t selectionTick = 'X'; //!<Selected element tick
-        int32_t progress = (13 * 16) + 11; //!<Progress bar fill
+        uint8_t borderHorizontal = '-'; //!<Horizontal border character
+        uint8_t borderVertical = '|'; //!<Vertical border character
+        uint8_t borderCornerTl = '+'; //!<Top left corner border character
+        uint8_t borderCornerTr = '+'; //!<Top right corner border character
+        uint8_t borderCornerBl = '+'; //!<Bottom left corner border character
+        uint8_t borderCornerBr = '+'; //!<Bottom right corner border character
+        uint8_t scrollbarCharacter = (11 * 16) + 10; //!<Scrollbar character
+        uint8_t scrollbarCharacterHorizontal = (12 * 16) + 13; //!<Horizontal scrollbar character
+        uint8_t selectionTick = 'X'; //!<Selected element tick
+        uint8_t progress = (13 * 16) + 11; //!<Progress bar fill
 
         SDL_Color background = { 0, 0, 0, 255 }; //!<Background for widgets
         SDL_Color text = { 255, 255, 255, 255 }; //!<Text for widgets
